Import multiple pictures. Handle gray scale imports.
Moved some actions to a menu.
It is now possible to import multiple face pictures of the same person
by selecting them in the import file chooser. Also the profile picture
may be imported separately.

A new feature is to be able to display the person which is the closest
look alike :-)